+++ packages/ui/src/cairo/AccessControlSection.svelte
@@ -36,7 +36,7 @@
       <span class="ml-1">
         <ToggleRadio bind:value={access} defaultValue="ownable" disabled={required} />
       </span>
-      <HelpTooltip align="right" link="https://github.com/OpenZeppelin/cairo-contracts/blob/main/docs/Access.md">
+      <HelpTooltip align="right" link="https://docs.openzeppelin.com/contracts-cairo/access">
         Restrict who can access the functions of a contract or when they can do it.
       </HelpTooltip>
     </label>
@@ -46,14 +46,14 @@
     <label class:checked={access === 'ownable'}>
       <input type="radio" bind:group={access} value="ownable">
       Ownable
-      <HelpTooltip link="https://github.com/OpenZeppelin/cairo-contracts/blob/main/docs/Access.md#ownable">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/access#ownable">
         Simple mechanism with a single account authorized for all privileged actions.
       </HelpTooltip>
     </label>
     <label class:checked={access === 'roles'}>
       <input type="radio" bind:group={access} value="roles">
       Roles
-      <HelpTooltip link="https://github.com/OpenZeppelin/cairo-contracts/blob/main/docs/Access.md#accesscontrol">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/access#accesscontrol">
         Flexible mechanism with a separate role for each privileged action. A role can have many authorized accounts.
       </HelpTooltip>
     </label>