+++ packages/ui/src/cairo/ERC20Controls.svelte
@@ -59,7 +59,7 @@
     <label class:checked={opts.mintable}>
       <input type="checkbox" bind:checked={opts.mintable}>
       Mintable
-      <HelpTooltip link="https://github.com/OpenZeppelin/cairo-contracts/blob/main/docs/ERC20.md#erc20_mintable">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/erc20#erc20mintable">
         Privileged accounts will be able to create more supply.
       </HelpTooltip>
     </label>
@@ -75,7 +75,7 @@
     <label class:checked={opts.pausable}>
       <input type="checkbox" bind:checked={opts.pausable}>
       Pausable
-      <HelpTooltip link="https://github.com/OpenZeppelin/cairo-contracts/blob/main/docs/Security.md#pausable">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/security#pausable">
         Privileged accounts will be able to pause the functionality marked with <code>assert_not_paused</code>.
         Useful for emergency response.
       </HelpTooltip>