+++ packages/core-cairo/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.5.0 (2022-09-22)
+
+- Add `owner` view for Ownable. ([#165](https://github.com/OpenZeppelin/contracts-wizard/issues/165))
+- Implement Cairo 0.10 syntax changes. ([#173](https://github.com/OpenZeppelin/contracts-wizard/issues/173))
+
 ## 0.4.0 (2022-08-04)
 
 - Update directory structure for libraries. ([#157](https://github.com/OpenZeppelin/contracts-wizard/pull/157))